feat : 버스 위치 정보 가공 기능 구현

+++ server/src/main/java/com/talkka/server/bus/dao/BusLocationEntity.java
@@ -42,7 +42,7 @@ public class BusLocationEntity {
 	private String apiStationId;
 
 	@Column(name = "station_seq", nullable = false)
-	private Short stationSeq;
+	private Integer stationSeq;
 
 	@Column(name = "end_bus", nullable = false, length = 1)
 	@Convert(converter = EndBusConverter.class)
@@ -60,7 +60,7 @@ public class BusLocationEntity {
 	private PlateType plateType;
 
 	@Column(name = "remain_seat_count", nullable = false)
-	private Short remainSeatCount;
+	private Integer remainSeatCount;
 
 	@Column(name = "api_call_no", nullable = false)
 	private Integer apiCallNo;
